+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## Unreleased
-- fix: run original function in runInAsyncContext/runPromiseInAsyncContext when tracing is not active.
+## 1.114.0
 - Introduction of AWS SQS instrumentation for sending and reading messages.
+- fix: run original function in runInAsyncContext/runPromiseInAsyncContext when tracing is not active.
 
 ## 1.113.0
 - Instrument MongoDB native driver unified topology.